+++ src/component/DefaultTooltipContent.js
@@ -50,8 +50,12 @@ class DefaultTooltipContent extends Component {
     if (payload && payload.length) {
       const listStyle = { padding: 0, margin: 0 };
 
-      const items = payload.sort(itemSorter)
+      const items = (itemSorter ? _.sortBy(payload, itemSorter) : payload)
         .map((entry, i) => {
+          if (entry.type === 'none') {
+            return null;
+          }
+
           const finalItemStyle = {
             display: 'block',
             paddingTop: 4,
@@ -59,20 +63,21 @@ class DefaultTooltipContent extends Component {
             color: entry.color || '#000',
             ...itemStyle,
           };
-          const hasName = isNumOrStr(entry.name);
           const finalFormatter = entry.formatter || formatter || defaultFormatter;
-
+          let { name, value } = entry;
+          if (finalFormatter) {
+            const formatted = finalFormatter(value, name, entry, i);
+            if (Array.isArray(formatted)) {
+              [value, name] = formatted;
+            } else {
+              value = formatted;
+            }
+          }
           return (
             <li className="recharts-tooltip-item" key={`tooltip-item-${i}`} style={finalItemStyle}>
-              {hasName ? <span className="recharts-tooltip-item-name">{entry.name}</span> : null}
-              {
-                hasName ?
-                  <span className="recharts-tooltip-item-separator">{separator}</span> :
-                  null
-              }
-              <span className="recharts-tooltip-item-value">
-                {finalFormatter ? finalFormatter(entry.value, entry.name, entry, i) : entry.value}
-              </span>
+              {isNumOrStr(name) ? <span className="recharts-tooltip-item-name">{name}</span> : null}
+              {isNumOrStr(name) ? <span className="recharts-tooltip-item-separator">{separator}</span> : null}
+              <span className="recharts-tooltip-item-value">{value}</span>
               <span className="recharts-tooltip-item-unit">{entry.unit || ''}</span>
             </li>
           );